requirements.txt
#python==3.8
pandas==1.4.4
seaborn
spacy==3.7.2
typing_extensions==4.4.0
tqdm==4.64.1
natsort
scikit-learn>=1.0.0
nltk==3.7
# spacy-transformers
# python -m spacy download en_core_web_trf
# python -m spacy download en_core_web_sm
# en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl
transformers==4.26.1
torch==1.13.1

# backtranslation-related modules
sentence-transformers #for semantic simarlity of backtranslated review with original review
simalign==0.3         #for aligning words of backtranslated review to the original review
# networkx==3.1
rouge                 #for evaluating backtranlated reviews vs. original review

# topic modeling methods for aspect modeling
gensim==4.*
bitermplus==0.6.10          #install it if you want to use btm model in the aspect modeling layer
contextualized_topic_models==2.5.0 #install it if you want to use ctm model in the aspect modeling layer
# octis # #use Fani's Lab! fork at https://github.com/fani-lab/OCTIS followed by "cd src/octis/; python setup.py install"

pytrec-eval-terrier

bert-e2e-absa @ git+https://github.com/fani-lab/BERT-E2E-ABSA.git@master

sentencepiece == 0.1.99
deep-translator == 1.11.4

fasttext @ git+https://github.com/facebookresearch/fastText.git

pampy==0.3.0
returns==0.22.0
more-itertools==10.1.0

# needed for implicit dataset generation
openai
google-generativeai
python-dotenv